# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TRKR Invoice Tracking API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /suppliers:
    get:
      operationId: listSuppliers
      tags: [suppliers]
      summary: List suppliers
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: List of suppliers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supplier"
    post:
      operationId: createSupplier
      tags: [suppliers]
      summary: Create a supplier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSupplierBody"
      responses:
        "201":
          description: Created supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
  /suppliers/{id}:
    get:
      operationId: getSupplier
      tags: [suppliers]
      summary: Get supplier by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
    patch:
      operationId: updateSupplier
      tags: [suppliers]
      summary: Update supplier
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSupplierBody"
      responses:
        "200":
          description: Updated supplier
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supplier"
...
  /staff:
    get:
      operationId: listStaff
      tags: [staff]
      summary: List staff routing rules
      responses:
        "200":
          description: List of staff
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/StaffRoute"
    post:
      operationId: createStaff
      tags: [staff]
      summary: Create a staff routing rule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffBody"
      responses:
        "201":
          description: Created staff entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRoute"
  /staff/{id}:
    patch:
      operationId: updateStaff
      tags: [staff]
      summary: Update staff routing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateStaffBody"
      responses:
        "200":
          description: Updated staff entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffRoute"
    delete:
      operationId: deleteStaff
      tags: [staff]
      summary: Delete staff routing rule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required: [message]
...
    Supplier:
      type: object
      properties:
        id:
          type: integer
        supplierID:
          type: string
        supplierName:
          type: string
        fiCalID:
          type: string
          nullable: true
        sb:
          type: boolean
        mb:
          type: boolean
        dvbe:
          type: boolean
        createdAt:
          type: string
      required: [id, supplierID, supplierName, sb, mb, dvbe, createdAt]
    CreateSupplierBody:
      type: object
      properties:
        supplierID:
          type: string
        supplierName:
          type: string
        fiCalID:
          type: string
          nullable: true
        sb:
          type: boolean
        mb:
          type: boolean
        dvbe:
          type: boolean
      required: [supplierID, supplierName]
    UpdateSupplierBody:
      type: object
      properties:
        supplierName:
          type: string
          nullable: true
        fiCalID:
          type: string
          nullable: true
        sb:
          type: boolean
          nullable: true
        mb:
          type: boolean
          nullable: true
        dvbe:
          type: boolean
          nullable: true
...
    StaffRoute:
      type: object
      properties:
        id:
          type: integer
        reportingStructure:
          type: string
        invoiceType:
          type: string
          nullable: true
        accountant:
          type: string
        supervisor:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
        section:
          type: string
          nullable: true
        group:
          type: string
          nullable: true
        divisionCenter:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, reportingStructure, accountant, createdAt]
    CreateStaffBody:
      type: object
      properties:
        reportingStructure:
          type: string
        invoiceType:
          type: string
          nullable: true
        accountant:
          type: string
        supervisor:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
        section:
          type: string
          nullable: true
        group:
          type: string
          nullable: true
        divisionCenter:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
      required: [reportingStructure, accountant]
    UpdateStaffBody:
      type: object
      properties:
        reportingStructure:
          type: string
          nullable: true
        invoiceType:
          type: string
          nullable: true
        accountant:
          type: string
          nullable: true
        supervisor:
          type: string
          nullable: true
        unit:
          type: string
          nullable: true
        branch:
          type: string
          nullable: true
        section:
          type: string
          nullable: true
        group:
          type: string
          nullable: true
        divisionCenter:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
...