# lib/api-zod/src/generated/types/staffRoute.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRKR Invoice Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface StaffRoute {
  id: number;
  reportingStructure: string;
  invoiceType?: string | null;
  accountant: string;
  supervisor?: string | null;
  unit?: string | null;
  branch?: string | null;
  section?: string | null;
  group?: string | null;
  divisionCenter?: string | null;
  description?: string | null;
  createdAt: string;
}