# lib/api-zod/src/generated/types/updateStaffBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRKR Invoice Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateStaffBody {
  reportingStructure?: string | null;
  invoiceType?: string | null;
  accountant?: string | null;
  supervisor?: string | null;
  unit?: string | null;
  branch?: string | null;
  section?: string | null;
  group?: string | null;
  divisionCenter?: string | null;
  description?: string | null;
}