# lib/api-zod/src/generated/types/updateSupplierBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TRKR Invoice Tracking API
 * OpenAPI spec version: 0.1.0
 */

export interface UpdateSupplierBody {
  supplierName?: string | null;
  fiCalID?: string | null;
  sb?: boolean | null;
  mb?: boolean | null;
  dvbe?: boolean | null;
}